requirements.txt
Django==4.2.2
ipdb==0.13.13
flake8==6.0.0
django-environ==0.10.0
Pillow==10.0.0
Faker==19.2.0
pylint-django==2.5.3
bleach==6.0.0
django-widget-tweaks==1.4.12
weasyprint==59.0
django-lifecycle==1.0.0
celery[redis]==5.3.1
django-celery-results==2.5.1
django-model-utils==4.3.1
django_extensions==3.2.3
django_celery_beat==2.5.0
django-memcached==0.1.2
django-pymemcache==1.0.0
django-rosetta==0.9.9
django-silk==5.0.3
django-debug-toolbar==4.1.0
pytest==7.4.0
pytest-django==4.5.2
pytest-cov==4.1.0
PyPDF2==3.0.1
pytest-factoryboy==2.5.1

